+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -984,7 +984,7 @@ export class WorkflowService {
 
     return updatedWorkflow;
   }
-  
+
   async updateContextById(id: string, context: WorkflowRuntimeData['context']) {
     return this.workflowRuntimeDataRepository.updateContextById(id, context);
   }