+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ballerine/workflows-service
 
+## 0.4.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.18
+  - @ballerine/workflow-core@0.4.28
+  - @ballerine/workflow-node-sdk@0.4.15
+
 ## 0.4.28
 
 ### Patch Changes